// src/api/models/UpdatePlayer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdatePlayer
 */
export interface UpdatePlayer {
    /**
     * 
     * @type {string}
     * @memberof UpdatePlayer
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdatePlayer
     */
    lastName?: string;
    /**
     * 
     * @type {Date}
     * @memberof UpdatePlayer
     */
    birthDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof UpdatePlayer
     */
    firstName?: string;
}

/**
 * Check if a given object implements the UpdatePlayer interface.
 */
export function instanceOfUpdatePlayer(value: object): boolean {
    return true;
}

export function UpdatePlayerFromJSON(json: any): UpdatePlayer {
    return UpdatePlayerFromJSONTyped(json, false);
}

export function UpdatePlayerFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdatePlayer {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'] == null ? undefined : json['email'],
        'lastName': json['lastName'] == null ? undefined : json['lastName'],
        'birthDate': json['birthDate'] == null ? undefined : (new Date(json['birthDate'])),
        'firstName': json['firstName'] == null ? undefined : json['firstName'],
    };
}

export function UpdatePlayerToJSON(value?: UpdatePlayer | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'email': value['email'],
        'lastName': value['lastName'],
        'birthDate': value['birthDate'] == null ? undefined : ((value['birthDate']).toISOString().substring(0,10)),
        'firstName': value['firstName'],
    };
}